requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.29.0
streamlit-chat>=0.1.1

# RAG and LLM Framework
llama-index>=0.9.20
openai>=1.6.0
python-dotenv>=1.0.0

# Vector Database
pinecone>=3.0.0

# Traditional Database
pymongo>=4.6.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-magic-bin>=0.4.14  # For Windows compatibility

# Data Handling and Analysis
pandas>=2.1.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
streamlit-plotly-events>=0.0.6

# HTTP and API
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.2
pydantic>=2.5.0
typing-extensions>=4.8.0

# Development and Testing
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0

# Text Processing
langdetect>=1.0.9
beautifulsoup4>=4.12.0

# Async Support
asyncio>=3.4.3

# Logging and Monitoring
loguru>=0.7.0

# Security
cryptography>=41.0.0 